==> 1. Data Preprocessing/2. Data Preparation/1. Data rounding and cleansing.ipynb <==
# Data Rounding and Cleansing

import pandas as pd
import numpy as np

## 1. Load the sampled data

patients = pd.read_csv('../../Data/complete_dataset_zero_missings.csv', sep=",")
patients.drop(columns=["Unnamed: 0"], inplace=True)
patients.head()

#patients = pd.read_csv("../Data/total_patients_balanced.csv",index_col=0, header=0)
#patients.dropna(inplace=True)
#patients.head()

## 2. Removing the unnecessary features

unnecessary_features = ['ethnic', 'BMI', 'BMI_category', 'Waist_category', 'waist', 'dose_pleasure','dose_sport','dose_execise','dose_lightDIY','dose_heavyDIY','Alcohol24']

patients.drop(unnecessary_features, axis=1, inplace=True)

patients.shape, patients.columns.values

patients.reset_index(inplace=True)

patients.index.values

patients_integer = patients.round(0)[['Class','sex','age','HIPX','menopause', 'HRT','smoking', 'ReumatoidArthritis', 'SecondaryOsteoporsis']] #'ethnic','BMI_category', 'Waist_category',
patients_float = patients.round(2)[['weight', 'height', 'Alcohol','VitaminD', 'calcium', 'dose_walk', 'dose_moderate','dose_vigorous']] #'BMI',,'Alcohol24', 'dose_plea'waist',sure', 'dose_sport', 'dose_execise','dose_lightDIY', 'dose_heavyDIY'
#patients = pd.concat([patients_integer,patients_float])

put the two dataframe together, then reorder the columns as the original dataset, and then check that every value in the original dataframe and the rounded dataframe are equals, to ensure that no data has been changed.

rounded_patients = patients_integer.join(patients_float)

patients.drop(['index'], axis=1, inplace=True)

rounded_patients.shape == patients.shape

rounded_patients = rounded_patients[patients.columns.tolist()]

patients.columns.tolist() == rounded_patients.columns.tolist()

for i in range(0,rounded_patients.shape[0]):
    results = rounded_patients.iloc[i] == patients.iloc[i].round(2)
    for key, value in results.to_dict().items() :
        if not value :
            if key not in patients_integer.columns.values:
                print("For the patients with index: ",i," there is a difference in: ", key)

## Randomize data order

print("Dataframe shape before sorting: ",rounded_patients.shape)
rounded_patients = rounded_patients.sample(frac=1).reset_index(drop=True)
print("Dataframe shape after sorting: ",rounded_patients.shape)
rounded_patients.head()

rounded_patients.to_csv("../../Data/rounded_patients.csv")
