==> 1. Data Preprocessing/2. Data Preparation/2. Data Standardization ( no-categorical ).ipynb <==
# Data Standardization

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler

## 1. Load the Data

patients = pd.read_csv("../Data/rounded_patients.csv", index_col=0)

patients.head()

## 2. Standardize the patients

### 2.1 Remove the categorical variable from the Data

patients.columns

categorical_features = ['sex','HIPX','menopause','HRT','smoking', 'ReumatoidArthritis', 
                        'SecondaryOsteoporsis', 'Alcohol24', 'Class']

std_patients = patients.drop(labels=categorical_features, axis=1)
std_patients.shape

std_patients = patients

### 2.2 Standardize the features

std_patients = pd.DataFrame(StandardScaler().fit_transform(std_patients.iloc[:,0:std_patients.shape[1]-1]), columns=std_patients.columns.values[:std_patients.shape[1]-1])

## 3. Append the classification column and all the categorical features

for feature in categorical_features:
    std_patients[feature] = patients[feature]

std_patients['Class'] = patients['Class']

std_patients

## 4. Save to File

std_patients.to_csv("../Data/standardized_patients.csv")
