==> 1. Data Preprocessing/2. Data Preparation/3.2. Data_splitting (Data_set via t-SNE and Train, Test, and Validation).ipynb <==
# Dataset cleansing

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

## 1. Load the (standardized) Data

patients = pd.read_csv("../../Data/standardized_patients.csv",index_col=0, header=0)
patients.dropna(inplace=True)
patients.head()

patients = patients.sample(frac=1).reset_index(drop=True)

## 2. Create the two dataset from the t-SNE analysis

### 2.1 Women in menopause condition

women_menopause = patients[patients['menopause']>1]

women_menopause.drop(columns=['sex','menopause'], axis=1, inplace=True)

women_menopause.head()

#### Create the X and y

women_X = women_menopause.copy()
women_X.drop("Class", axis=1, inplace=True)
print(women_X.shape)

women_y = women_menopause[['Class']].copy()
print(women_y.shape)

#### Split the data into Train, Test and Validation

The ratio in the splitting, in percentage, is 60/20/20

women_X_train_and_val, women_X_test, women_y_train_and_val, women_y_test = train_test_split(women_X, women_y, test_size=0.2, random_state=42)
women_X_train, women_X_val, women_y_train, women_y_val = train_test_split(women_X_train_and_val, women_y_train_and_val, test_size=0.25, random_state=42)

women_y_val[women_y_val['Class']==1].shape

women_y_train[women_y_train['Class']==1].shape

women_y_test[women_y_test['Class']==1].shape

#### Save to Files

patients.to_csv("../../Data/women_menopause/standardized_patients.csv")

women_X_train.to_csv("../../Data/women_menopause/X_train_total.csv")
women_y_train.to_csv("../../Data/women_menopause/y_train_total.csv")

women_X_test.to_csv("../../Data/women_menopause/X_test.csv")
women_y_test.to_csv("../../Data/women_menopause/y_test.csv")

women_X_val.to_csv("../../Data/women_menopause/X_val.csv")
women_y_val.to_csv("../../Data/women_menopause/y_val.csv")

### 2.2 Other patients

other_patients = patients[patients['menopause']<1]

other_patients.drop(columns=['menopause','HRT'], axis=1, inplace=True)

other_patients.head()

#### Create the X and y

other_patients_X = other_patients.copy()
other_patients_X.drop("Class", axis=1, inplace=True)
print(other_patients.shape)

other_patients_y = other_patients[['Class']].copy()
print(other_patients_y.shape)

#### Split the data into Train, Test and Validation

The ratio in the splitting, in percentage, is 60/20/20

other_patients_X_train_and_val, other_patients_X_test, other_patients_y_train_and_val, other_patients_y_test = train_test_split(other_patients_X, other_patients_y, test_size=0.2, random_state=42)
other_patients_X_train, other_patients_X_val, other_patients_y_train, other_patients_y_val = train_test_split(other_patients_X_train_and_val, other_patients_y_train_and_val, test_size=0.25, random_state=42)

other_patients_y_val[other_patients_y_val['Class']==1].shape

other_patients_y_train[other_patients_y_train['Class']==1].shape

other_patients_y_test[other_patients_y_test['Class']==1].shape

#### Save to Files

patients.to_csv("../../Data/other_patients/standardized_patients.csv")

other_patients_X_train.to_csv("../../Data/other_patients/X_train_total.csv")
other_patients_y_train.to_csv("../../Data/other_patients/y_train_total.csv")

other_patients_X_test.to_csv("../../Data/other_patients/X_test.csv")
other_patients_y_test.to_csv("../../Data/other_patients/y_test.csv")

other_patients_X_val.to_csv("../../Data/other_patients/X_val.csv")
other_patients_y_val.to_csv("../../Data/other_patients/y_val.csv")
