==> 1. Data Preprocessing/2. Data Preparation/4.3. Women_Menopause Data balancing (Random Undersampling and Oversampling).ipynb <==
# Data Balancing - Random undersampling class 0 and oversampling class 1

import pandas as pd
import numpy as np
from collections import Counter
from imblearn.over_sampling import ADASYN
from imblearn.under_sampling import RandomUnderSampler

## 1. Load the Data

starting_ratio = ''#'starting_ratio_05/'
percent_ratio = ''# '10_percent/'
base_data_folder = '../../Data/women_menopause/'+starting_ratio+percent_ratio

X_train = pd.read_csv(base_data_folder+'X_train_total.csv', index_col=0, header=0)
y_train = pd.read_csv(base_data_folder+'y_train_total.csv', index_col=0, header=0)
X_train.shape, y_train.shape

## 2. Undersampling

ratio = 1

y_train[y_train['Class'] == 1].shape

rus = RandomUnderSampler(sampling_strategy=ratio, random_state=42)
X_resampled, y_resampled = rus.fit_resample(X_train, y_train['Class'].values)
print('Resampled dataset shape %s' % Counter(y_resampled))

X_resampled.shape

y_resampled.shape

## 3. Oversampling

from imblearn.over_sampling import ADASYN

adasyn = ADASYN(random_state=42, sampling_strategy=ratio)
X_resampled_total, y_resampled_total = adasyn.fit_resample(X_resampled, y_resampled)
print('Resampled dataset shape %s' % Counter(y_resampled_total))

total_patients = pd.DataFrame(X_resampled_total)
total_patients["Class"] = y_resampled_total

## 4. Dataset merge and Data preparation for disk writing

total_patients = pd.DataFrame(X_resampled_total)
total_patients["Class"] = y_resampled_total

columns_names = ['age', 'weight', 'height', 'HIPX', 'HRT', 'smoking',
       'ReumatoidArthritis', 'SecondaryOsteoporsis', 'Alcohol', 'VitaminD',
       'calcium', 'dose_walk', 'dose_moderate', 'dose_vigorous','Class']

total_patients.columns = columns_names

total_patients = total_patients.sample(frac=1).reset_index(drop=True)

X_train_resampled = total_patients.iloc[:,:total_patients.shape[1]-1]
X_train_resampled

y_train_resampled = pd.DataFrame(total_patients['Class'], columns=['Class'])
y_train_resampled

## 5. Save to file

ratio_folder = ''
if starting_ratio == '':
    if ratio < 1 :
        base_data_folder += 'starting_ratio_05/'
    else:
        base_data_folder += 'starting_ratio_1/'
else:
    if ratio < 1 :
        ratio_folder = 'ratio_05/'
    else:
        ratio_folder = 'ratio_1/'
        
total_patients.to_csv(base_data_folder+ratio_folder+'total_patients_trainset_balanced.csv')
X_train_resampled.to_csv(base_data_folder+ratio_folder+'X_train.csv')
y_train_resampled.to_csv(base_data_folder+ratio_folder+'y_train.csv')

for j in range (1,6):
    print(j)
    starting_ratio = '1'
    percent_ratio = str(j)+'0'
    base_data_folder = '../../Data/women_menopause/starting_ratio_'+starting_ratio+'/'+percent_ratio+'_percent/'
    
    X_train = pd.read_csv(base_data_folder+'X_train_total.csv', index_col=0, header=0)
    y_train = pd.read_csv(base_data_folder+'y_train_total.csv', index_col=0, header=0)
    X_train.shape, y_train.shape
    
    for i in [0.5,1]:
        
        ratio = i

        rus = RandomUnderSampler(sampling_strategy=ratio, random_state=42)
        X_resampled, y_resampled = rus.fit_resample(X_train, y_train['Class'].values)

        adasyn = ADASYN(random_state=42, sampling_strategy=ratio)
        X_resampled_total, y_resampled_total = adasyn.fit_resample(X_resampled, y_resampled)

        total_patients = pd.DataFrame(X_resampled_total)
        total_patients["Class"] = y_resampled_total

        columns_names = ['age', 'weight', 'height', 'HIPX', 'HRT', 'smoking',
       'ReumatoidArthritis', 'SecondaryOsteoporsis', 'Alcohol', 'VitaminD',
       'calcium', 'dose_walk', 'dose_moderate', 'dose_vigorous','Class']
        total_patients.columns = columns_names
        total_patients = total_patients.sample(frac=1).reset_index(drop=True)

        X_train_resampled = total_patients.iloc[:,:total_patients.shape[1]-1]

        y_train_resampled = pd.DataFrame(total_patients['Class'], columns=['Class'])

        if ratio < 1:
            ratio_folder = 'ratio_05/'
        else:
            ratio_folder = 'ratio_1/'
        total_patients.to_csv(base_data_folder+ratio_folder+'total_patients_trainset_balanced.csv')
        X_train_resampled.to_csv(base_data_folder+ratio_folder+'X_train.csv')
        y_train_resampled.to_csv(base_data_folder+ratio_folder+'y_train.csv')
