==> 1. Data Preprocessing/2. Data Preparation/5. Matlab data preparation.ipynb <==
# Matlab data preparation

import pandas as pd
import numpy as np

## 1. Load the standardized data

input_patients = pd.read_csv('../Data/X_train.csv', index_col=0, header=0)
target_patients = pd.read_csv('../Data/y_train.csv', index_col=0, header=0)

## 3. Save to files

input_patients.to_csv("../Data/matlab_input_patients.csv", index=False)
target_patients.to_csv("../Data/matlab_target_patients.csv", index=False)
